+++ tests/testthat/test-appMetadata.R
@@ -1,6 +1,8 @@
 # appMetadata -------------------------------------------------------------
 
 test_that("quarto affects mode inference", {
+  skip_on_cran()
+
   dir <- local_temp_app(list("foo.Rmd" = ""))
 
   metadata <- appMetadata(dir, c("foo.Rmd"))
@@ -11,12 +13,16 @@ test_that("quarto affects mode inference", {
 })
 
 test_that("quarto path is deprecated", {
+  skip_on_cran()
   skip_if_no_quarto()
+
   dir <- local_temp_app(list("foo.Rmd" = ""))
   expect_snapshot(. <- appMetadata(dir, c("foo.Rmd"), quarto = "abc"))
 })
 
 test_that("validates quarto argument", {
+  skip_on_cran()
+
   dir <- local_temp_app(list("foo.Rmd" = ""))
   expect_snapshot(appMetadata(dir, c("foo.Rmd"), quarto = 1), error = TRUE)
 })