+++ tests/testthat/test-http-internal.R
@@ -1,4 +1,6 @@
 test_that("basic HTTP methods work", {
+  skip_if_not_installed("webfakes")
+
   withr::local_options(rsconnect.http = "internal", lifecycle_verbosity = "quiet")
 
   test_http_GET()